# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API",
        "description": "Documentação apresentando todos endpoints desta API. Contém informações de autenticação, corpo da requisição e respostas.",
        "contact": {
            "email": "hiagolaureano@hotmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
...
        "/v1/pets": {
            "get": {
                "tags": [
                    "/v1/pets"
                ],
                "summary": "Obter lista de pets",
                "description": "Retorna uma lista com todos pets",
                "operationId": "1dc316bd42f20e9cb638b9ff896a5207",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da paginação dos dados. São 15 registros por paginação",
                        "style": "form",
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Dados da requisição",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Pet"
                                            }
                                        },
                                        "links": {
                                            "description": "Links da paginação",
                                            "type": "object",
                                            "example": {
                                                "first": "http:localhost/api/v1/item?page=1",
                                                "last": "http:localhost/api/v1/item?page=3",
                                                "prev": "http:localhost/api/v1/item?page=1",
                                                "next": "http:localhost/api/v1/item?page=3"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Login necessário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Somente superusuários possuem essa permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "/v1/pets"
                ],
                "summary": "Registrar novo pet",
                "description": "Cria um novo pet",
                "operationId": "62c0a1366755b71f4c9ce2a350ec52c8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "species",
                                    "breed",
                                    "weight",
                                    "gender",
                                    "agressive"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "Id do usuário dono, se não for fornecido, será atribuído o usuário logado",
                                        "type": "integer"
                                    },
                                    "name": {
                                        "description": "Nome",
                                        "type": "string"
                                    },
                                    "species": {
                                        "description": "Espécie",
                                        "type": "string"
                                    },
                                    "breed": {
                                        "description": "Raça",
                                        "type": "string"
                                    },
                                    "weight": {
                                        "description": "Peso",
                                        "type": "numeric"
                                    },
                                    "age": {
                                        "description": "Idade",
                                        "type": "integer"
                                    },
                                    "gender": {
                                        "description": "Sexo, 'M' para macho e 'F' para fêmea",
                                        "type": "string"
                                    },
                                    "agressive": {
                                        "description": "Se é agressivo",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Novo registro criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pet"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Login necessário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Somente o propretário do registro e superusuários possuem essa permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "420": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "pattern": "O campo x é obrigatório. (+ 3 erros)"
                                        },
                                        "errors": {
                                            "description": "Campos com erro de validação e lista dos erros",
                                            "type": "object",
                                            "example": {
                                                "campo_x": [
                                                    "O campo x é obrigatório."
                                                ],
                                                "campo_y": [
                                                    "O campo y deve ser um inteiro."
                                                ],
                                                "campo_z": [
                                                    "O campo z deve ter no máximo 30 caracteres."
                                                ],
                                                "campo_w": [
                                                    "O campo w é obrigatório."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Pet": {
                "title": "Pet",
                "description": "Pet Model",
                "properties": {
                    "id": {
                        "description": "Id do pet",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "Id do usuário dono do pet",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nome",
                        "type": "string"
                    },
                    "species": {
                        "description": "Espécie",
                        "type": "string"
                    },
                    "breed": {
                        "description": "Raça",
                        "type": "string"
                    },
                    "weight": {
                        "description": "Peso",
                        "type": "number"
                    },
                    "age": {
                        "description": "Idade",
                        "type": "integer"
                    },
                    "gender": {
                        "description": "Sexo, 'M' para macho e 'F' para fêmea",
                        "type": "string",
                        "pattern": "M"
                    },
                    "agressive": {
                        "description": "Se é agressivo",
                        "type": "boolean"
                    },
                    "created_at": {
                        "description": "Data da criação do registro",
                        "type": "string",
                        "pattern": "20/12/2024 11:12:40"
                    },
                    "updated_at": {
                        "description": "Data da atualização do registro",
                        "type": "string",
                        "pattern": "21/12/2024 11:15:30"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}